Added a Result page that will be automatically opened after a blast to show the analysis.
Added an "About this analysis" tab in the result page

Modified download buttons and possible download types

Cleaned the codebase

Styled a bit the pages

Added a description in the table in the result page to explain how to work with it. It uses st.expander so it can be shrunk to reduce the clutter in the page

Added some advanced option in the blast search, but still need to be customized depending on the blast algorithm and I need to implement them in the actual search. FOr now they are not used.

Cleaned the sidebar